Move parcel to types crate
