+++ core/src/block.rs
@@ -19,6 +19,7 @@ use std::collections::HashSet;
 use ccrypto::BLAKE_NULL_RLP;
 use ckey::Address;
 use cmerkle::skewed_merkle_root;
+use ctypes::parcel::Error as ParcelError;
 use primitives::{Bytes, H256};
 use rlp::{Decodable, DecoderError, Encodable, RlpStream, UntrustedRlp};
 use trie::TrieFactory;
@@ -30,7 +31,7 @@ use super::error::{BlockError, Error};
 use super::header::{Header, Seal};
 use super::invoice::Invoice;
 use super::machine::{LiveBlock, Parcels};
-use super::parcel::{ParcelError, SignedParcel, UnverifiedParcel};
+use super::parcel::{SignedParcel, UnverifiedParcel};
 use super::state::{ParcelOutcome, StateWithCache, TopLevelState};
 use super::state_db::StateDB;
