+++ core/src/codechain_machine.rs
@@ -16,13 +16,14 @@
 // A state machine.
 
 use ckey::Address;
+use ctypes::parcel::Error as ParcelError;
 use primitives::U256;
 
 use super::block::{ExecutedBlock, IsBlock};
 use super::client::BlockInfo;
 use super::error::Error;
 use super::header::Header;
-use super::parcel::{ParcelError, SignedParcel, UnverifiedParcel};
+use super::parcel::{SignedParcel, UnverifiedParcel};
 use super::spec::CommonParams;
 use super::state::{TopState, TopStateInfo};
