+++ core/src/error.rs
@@ -18,6 +18,7 @@ use std::fmt;
 
 use cio::IoError;
 use ckey::{Address, Error as KeyError};
+use ctypes::parcel::Error as ParcelError;
 use ctypes::transaction::Error as TransactionError;
 use primitives::{H256, U256};
 use trie::TrieError;
@@ -27,7 +28,6 @@ use util_error::UtilError;
 use super::account_provider::SignError as AccountsError;
 use super::client::Error as ClientError;
 use super::consensus::EngineError;
-use super::parcel::ParcelError;
 use super::types::BlockNumber;
 
 #[derive(Debug, Clone, Copy, PartialEq)]
