+++ core/src/lib.rs
@@ -100,7 +100,7 @@ pub use error::{BlockImportError, Error, ImportError};
 pub use header::{Header, Seal};
 pub use invoice::Invoice;
 pub use miner::{Miner, MinerOptions, MinerService};
-pub use parcel::{parcel_error_message, Action, LocalizedParcel, Parcel, SignedParcel, UnverifiedParcel};
+pub use parcel::{LocalizedParcel, SignedParcel, UnverifiedParcel};
 pub use service::ClientService;
 pub use spec::Spec;
 pub use state::{Asset, AssetAddress, AssetScheme, AssetSchemeAddress, ShardStateInfo, TopStateInfo};
