+++ core/src/miner/mem_pool.rs
@@ -19,13 +19,14 @@ use std::cmp::Ordering;
 use std::collections::{BTreeSet, HashMap};
 
 use ckey::Address;
+use ctypes::parcel::{Action, Error as ParcelError};
 use heapsize::HeapSizeOf;
 use linked_hash_map::LinkedHashMap;
 use multimap::MultiMap;
 use primitives::{H256, U256};
 use table::Table;
 
-use super::super::parcel::{Action, ParcelError, SignedParcel};
+use super::super::parcel::SignedParcel;
 use super::super::types::BlockNumber;
 use super::local_parcels::{LocalParcelsList, Status as LocalParcelStatus};
 use super::ParcelImportResult;
@@ -1094,9 +1095,9 @@ pub mod test {
     use std::cmp::Ordering;
 
     use ckey::{Generator, Random};
+    use ctypes::parcel::Parcel;
     use ctypes::transaction::{AssetMintOutput, Transaction};
 
-    use super::super::super::Parcel;
     use super::*;
 
     #[test]
@@ -1122,7 +1123,7 @@ pub mod test {
             },
         };
         let keypair = Random.generate().unwrap();
-        let signed = parcel.sign(keypair.private());
+        let signed = SignedParcel::new_with_sign(parcel, keypair.private());
         let item = MemPoolItem::new(signed, ParcelOrigin::Local, 0, 0);
 
         assert_eq!(fee, item.cost());
@@ -1151,7 +1152,7 @@ pub mod test {
             },
         };
         let keypair = Random.generate().unwrap();
-        let signed = parcel.sign(keypair.private());
+        let signed = SignedParcel::new_with_sign(parcel, keypair.private());
         let item = MemPoolItem::new(signed, ParcelOrigin::Local, 0, 0);
 
         assert_eq!(fee, item.cost());
@@ -1189,7 +1190,7 @@ pub mod test {
             },
         };
         let keypair = Random.generate().unwrap();
-        let signed = parcel.sign(keypair.private());
+        let signed = SignedParcel::new_with_sign(parcel, keypair.private());
         let item = MemPoolItem::new(signed, ParcelOrigin::Local, 0, 0);
 
         assert_eq!(fee, item.cost());
@@ -1210,7 +1211,7 @@ pub mod test {
                 amount,
             },
         };
-        let signed = parcel.sign(keypair.private());
+        let signed = SignedParcel::new_with_sign(parcel, keypair.private());
         let item = MemPoolItem::new(signed, ParcelOrigin::Local, 0, 0);
 
         assert_eq!(fee + amount, item.cost());
