+++ types/Cargo.toml
@@ -6,6 +6,7 @@ authors = ["CodeChain Team <codechain@kodebox.io>"]
 [dependencies]
 codechain-crypto = { path = "../crypto" }
 codechain-key = { path = "../key" }
+heapsize = "0.4"
 primitives = { path = "../util/primitives" }
 rlp = { path = "../util/rlp" }
 rlp_derive = { path = "../util/rlp_derive" }
